Neutralize unsafe (non-local) destination params
This addresses a cover redirect vulnerability when the following is used as an
authentication URL:

http://domain.dev/auth/provider?destination=//google.com

Covert redirects (https://www.owasp.org/index.php/Open_redirect) are often used
in phishing scams by giving people a link that appears save but surreptitiously
redirects them to a different domain.

Once on that different domain (which can be designed to look like the original
domain), users can be asked their password and give it away because they don't
notice that they have been redirected to a different destination.

This is not a vulnerability in itself, but used in combination with a rather
simple phishing attack, this can be potentially dangerous.